Feat : 좋아요 추가

+++ src/main/java/com/example/wherebnb/service/RoomsService.java
@@ -2,15 +2,14 @@ package com.example.wherebnb.service;
 
 import com.example.wherebnb.dto.ResponseDto;
 import com.example.wherebnb.dto.RoomsRequestDto;
-import com.example.wherebnb.dto.RoomsResponseDto;
+import com.example.wherebnb.entity.Likes;
 import com.example.wherebnb.entity.Rooms;
 import com.example.wherebnb.entity.Users;
-import com.example.wherebnb.global.ApiException;
-import com.example.wherebnb.global.ExceptionEnum;
+import com.example.wherebnb.exception.ApiException;
+import com.example.wherebnb.exception.ExceptionEnum;
+import com.example.wherebnb.repository.LikesRepository;
 import com.example.wherebnb.repository.RoomsRepository;
-import com.example.wherebnb.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class RoomsService {
 
     private final RoomsRepository roomsRepository;
+    private final LikesRepository likesRepository;
 
     // 숙소 등록
     public ResponseDto roomInsert(RoomsRequestDto roomsRequestDto, Users user) {
@@ -31,17 +31,38 @@ public class RoomsService {
     public ResponseDto roomUpdate(Long id, RoomsRequestDto roomsRequestDto, Users user) {
         Rooms room = roomsRepository.findById(id).orElseThrow(  // 수정할 게시글 있는지 확인
                 () -> new ApiException(ExceptionEnum.NOT_FOUND_ROOM));
+
+        if (!room.getUser().equals(user))
+            return ResponseDto.setBadRequest("숙소 수정을 할 수 없습니다.", null);
+
         room.roomUpdate(roomsRequestDto);
         return ResponseDto.setSuccess("숙소 수정을 완료하였습니다.", null);
     }
 
-
     // 숙소 삭제
-    public ResponseDto roomDelete(Long id, Users users) {
+    public ResponseDto roomDelete(Long id, Users user) {
         Rooms room = roomsRepository.findById(id).orElseThrow( // 삭제할 게시글 있는지 확인
                 () -> new ApiException(ExceptionEnum.NOT_FOUND_ROOM));
 
-       roomsRepository.delete(room);
+        if (!room.getUser().equals(user))
+            return ResponseDto.setBadRequest("숙소 수정을 할 수 없습니다.", null);
+
+        roomsRepository.delete(room);
         return ResponseDto.setSuccess("숙소 삭제를 완료하였습니다.", null);
+    }
+
+    public ResponseDto roomLikes(Long id, Users user) {
+        Rooms room = roomsRepository.findById(id).orElseThrow(() -> new ApiException(ExceptionEnum.NOT_FOUND_ROOM));
+        boolean likeStatus = true;
+
+        if (likesRepository.existsByUserIdAndRoomsId(user.getId(), room.getId())) { // 이미 좋아요한 경우 취소
+            Likes likes = likesRepository.findByUserIdAndRoomsId(user.getId(), room.getId());
+            likesRepository.delete(likes);
+            likeStatus = false;
+        } else
+            likesRepository.save(new Likes(room, user));
+
+        room.updateLikes(likeStatus);
+        return ResponseDto.setSuccess("좋아요를 눌렀습니다.", likeStatus);
     }
 }
